correct prototypes

+++ bld/ui/c/uifar.c
@@ -123,9 +123,9 @@ extern void _backward(void);
 
 #define ATTR_FLIP_MASK      0x77
 
-intern void cdecl farfill( LP_PIXEL start, PIXEL fill, unsigned len, int snow )
+intern void cdecl farfill( LP_PIXEL start, PIXEL fill, size_t len, int snow )
 {
-    unsigned    i;
+    size_t      i;
 
 #if defined( _M_I86 )
     if( snow ) {
@@ -146,7 +146,7 @@ intern void cdecl farfill( LP_PIXEL start, PIXEL fill, unsigned len, int snow )
 }
 
 
-intern void cdecl farcopy( LP_PIXEL src, LP_PIXEL dst, unsigned len, int snow )
+intern void cdecl farcopy( LP_PIXEL src, LP_PIXEL dst, size_t len, int snow )
 {
 #if defined( _M_I86 )
     if( snow ) {
@@ -177,10 +177,10 @@ intern void cdecl farcopy( LP_PIXEL src, LP_PIXEL dst, unsigned len, int snow )
 }
 
 
-intern void cdecl farstring( LP_PIXEL start, int attr, unsigned len,
+intern void cdecl farstring( LP_PIXEL start, int attr, size_t len,
                                           int snow, LPC_STRING str )
 {
-    unsigned    i;
+    size_t      i;
     PIXEL       p;
 
     p.attr = attr;
@@ -213,9 +213,9 @@ intern void cdecl farstring( LP_PIXEL start, int attr, unsigned len,
     farfill( start, p, len-i, snow );
 }
 
-intern void cdecl farattrib( LP_PIXEL start, int attr, unsigned len, int snow )
+intern void cdecl farattrib( LP_PIXEL start, int attr, size_t len, int snow )
 {
-    unsigned    i;
+    size_t      i;
     PIXEL       p;
 
 #if defined( _M_I86 )
@@ -239,9 +239,9 @@ intern void cdecl farattrib( LP_PIXEL start, int attr, unsigned len, int snow )
     }
 }
 
-intern void cdecl farattrflip( LP_PIXEL start, unsigned len, int snow )
+intern void cdecl farattrflip( LP_PIXEL start, size_t len, int snow )
 {
-    unsigned    i;
+    size_t      i;
     PIXEL       p;
 
 #if defined( _M_I86 )
